refactor: remove deprecated old queue implementation and tests
